Move screen analytics to navigation service
