+++ src/NavigatorService.ts
@@ -1,8 +1,10 @@
-import { NavigationContainerRef, CommonActions, StackActions, Route } from '@react-navigation/native';
+import { NavigationContainerRef, CommonActions, StackActions, Route, NavigationState } from '@react-navigation/native';
 
 import { ScreenParamList } from '@covid/features/ScreenParamList';
+import Analytics from '@covid/core/Analytics';
 
 let navigation: NavigationContainerRef;
+let currentRouteName = '';
 
 function setContainer(navigationRef: NavigationContainerRef) {
   navigation = navigationRef;
@@ -29,10 +31,38 @@ function goBack() {
   navigation!.goBack();
 }
 
+function handleStateChange() {
+  const state = navigation!.getRootState();
+  if (!state) return;
+
+  const previousRouteName = currentRouteName;
+  const newRouteName = getCurrentRouteName(state);
+
+  if (newRouteName) {
+    if (previousRouteName !== newRouteName) {
+      Analytics.trackScreenView(newRouteName);
+    }
+    currentRouteName = newRouteName;
+  }
+}
+
+const getCurrentRouteName = (navigationState: NavigationState): string | null => {
+  if (!navigationState) return null;
+
+  const route = navigationState.routes[navigationState.index];
+  if (route.state) {
+    // Nested navigators
+    // @ts-ignore
+    return getCurrentRouteName(route.state);
+  }
+  return route.name;
+};
+
 export default {
   setContainer,
   navigate,
   reset,
   replace,
   goBack,
+  handleStateChange,
 };
